Moved log_command to app_logging module
