+++ bot/handlers/command_handlers.py
@@ -11,6 +11,7 @@ from telegram.error import BadRequest
 from telegram.ext import CallbackContext
 
 import app_logging
+from app_logging.handler_logging import log_command
 from bot.chat_type_accepted import group_only_command
 from localization.replies import (
     start_message_private, start_message_chat,
@@ -29,40 +30,6 @@ from sql.domain import *
 
 # Registering logger here
 logger: logging.Logger = app_logging.get_logger(__name__)
-
-
-def log_command(command_name: str = None):
-    """
-    Designed to be a decorator.
-    Decorated function HAS TO have two arguments:
-    :class:`telegram.Update` and :class:`telegram.CallbackContext`
-    \n
-    Logs a message with a command from a user with such information, as chat_type, chat_id, user_id and command args.
-
-    Args:
-        command_name: name of the command to be logged. If not passed will be logged the first word in received message.
-    """
-
-    def log_command_decorator_maker(command_handler: Callable[[Update, CallbackContext], Any]):
-        def log_command_wrapped(update: Update, context: CallbackContext):
-            chat_id = update.effective_chat.id
-            chat_name = update.effective_chat.title
-            user_id = update.effective_user.id
-            chat_type = update.effective_chat.type
-            args = ' '.join(context.args) if context.args is not None else update.effective_message.text
-            info = f"chat_type: '{chat_type}', " \
-                   f"chat_id: '{chat_id}', chat_name: '{chat_name}', user: '{user_id}', args: '{args}'"
-            _command_name = command_name if command_name is not None else update.effective_message.text.split(' ')[0]
-            if update.edited_message:
-                logger.info(f"{_command_name}: [{info}] edited")
-            else:
-                logger.info(f"{_command_name} [{info}]")
-
-            return command_handler(update, context)
-
-        return log_command_wrapped
-
-    return log_command_decorator_maker
 
 
 def __insert_queue_from_context(on_no_queue_log: str, on_not_exist_log: str, on_no_queue_reply: dict):
